move the processing of action filters to the workflow layer - this will keep all the delayed processing / enqueuing of workflows into a single place.
