+++ src/sentry/workflow_engine/processors/workflow.py
@@ -4,9 +4,17 @@ from collections import defaultdict
 import sentry_sdk
 
 from sentry import buffer
+from sentry.db.models.manager.base_query_set import BaseQuerySet
 from sentry.utils import json, metrics
-from sentry.workflow_engine.models import Detector, Workflow, WorkflowDataConditionGroup
-from sentry.workflow_engine.processors.action import evaluate_workflows_action_filters
+from sentry.workflow_engine.models import (
+    Action,
+    DataCondition,
+    DataConditionGroup,
+    Detector,
+    Workflow,
+    WorkflowDataConditionGroup,
+)
+from sentry.workflow_engine.processors.action import filter_recently_fired_workflow_actions
 from sentry.workflow_engine.processors.data_condition_group import evaluate_condition_group
 from sentry.workflow_engine.processors.detector import get_detector_by_event
 from sentry.workflow_engine.types import WorkflowJob
@@ -81,6 +89,40 @@ def evaluate_workflow_triggers(workflows: set[Workflow], job: WorkflowJob) -> se
         enqueue_workflows(workflows_to_enqueue, job)
 
     return triggered_workflows
+
+
+def evaluate_workflows_action_filters(
+    workflows: set[Workflow],
+    job: WorkflowJob,
+) -> BaseQuerySet[Action]:
+    filtered_action_groups: set[DataConditionGroup] = set()
+    enqueued_conditions: list[DataCondition] = []
+
+    # gets the list of the workflow ids, and then get the workflow_data_condition_groups for those workflows
+    workflow_ids = {workflow.id for workflow in workflows}
+
+    action_conditions = DataConditionGroup.objects.filter(
+        workflowdataconditiongroup__workflow_id__in=workflow_ids
+    ).distinct()
+
+    for action_condition in action_conditions:
+        evaluation, result, remaining_conditions = evaluate_condition_group(action_condition, job)
+
+        if evaluation:
+            if remaining_conditions:
+                # If there are remaining conditions for the action filter to evaluate,
+                # then return the list of conditions to enqueue
+                enqueued_conditions.extend(remaining_conditions)
+            else:
+                # if we don't have any other conditions to evaluate, add the action to the list
+                filtered_action_groups.add(action_condition)
+
+    # get the actions for any of the triggered data condition groups
+    actions = Action.objects.filter(
+        dataconditiongroupaction__condition_group__in=filtered_action_groups
+    ).distinct()
+
+    return filter_recently_fired_workflow_actions(actions, job["event"].group)
 
 
 def process_workflows(job: WorkflowJob) -> set[Workflow]:
